Make sampler length equal to the number of returned batches (#212) (#213)
* Make sampler length equal to the number of returned batches (#212)

* Add DataSource - manages data shuffling, remembers original order
* Add num_batches attribute and return correct number of batches
* Tests - check length of samplers, determinism, changing order each epoch

* Use list comprehension instead of list(sampler) in tests

* Return Long tensor from TokenCollater to comply with torch.nn.Embedding

* Fix audio and feature padding length in collation

* Adjust distributed sampler init to comply with torch1.8

+++ lhotse/dataset/speech_synthesis.py
@@ -17,7 +17,7 @@ class SpeechSynthesisDataset(torch.utils.data.Dataset):
         {
             'audio': (B x NumSamples) float tensor
             'features': (B x NumFrames x NumFeatures) float tensor
-            'tokens': (B x NumTokens) int tensor
+            'tokens': (B x NumTokens) long tensor
             'audio_lens': (B, ) int tensor
             'features_lens': (B, ) int tensor
             'tokens_lens': (B, ) int tensor
